requirements.txt
fastapi
uvicorn
httpx
pyyaml
pymodbus
mss
requests
Pillow
openai
qrcode
pycomm3
opcua
python-multipart
aiosqlite
opencv-python
python-dotenv
